public

+++ app/Http/Controllers/AdminController.php
@@ -10,10 +10,12 @@ use App\Shop;
 use App\Floor;
 
 use App\Image;
+use App\Slide;
 use Storage;
 
 class AdminController extends Controller
 {
+
     public function __construct()
     {
         $this->middleware('auth');
@@ -39,17 +41,45 @@ class AdminController extends Controller
         }
     }
 
+/*
+    public function multiple_slide_upload($model, $request)
+    {
+        if($model and $request->sliders) {
+            foreach ($request->sliders as $slide) {
+                 $filesize = $slide->getSize();
+                 $filename = md5(time().$filesize).'.'.$slide->getClientOriginalExtension();
+                 $destinationPath = 'storage/sliders';
+                 $upload_success = $slide->move($destinationPath, $filename);
+                 if($upload_success) {
+
+                        $slide = new Slider([
+                                'name' => $model->name,
+                                'filename' => $filename,
+                                'filesize' => $filesize,
+
+                            ]);
+                        $model->sliders()->save($slide);
+
+                 }
+            }
+        }
+    }
+*/
     public function index()
     {
        # return view('admin.admin');
     }
 
 
+
+
+
+
 /************** Магазины **************/
 
     public function index_store(Request $request)
     {
-        $stores = Store::with('shop');
+        $stores = Store::with('shop', 'avatar', 'floor');
         $stores = $stores->orderby('updated_at', 'desc')->paginate(20);
         return view('store.admin', compact('stores'));
     }
@@ -88,6 +118,12 @@ class AdminController extends Controller
     public function delete_store(Request $request)
     {
         $store = Store::find($request->id);
+        if($store->posts->count() >= 1) {
+         return redirect()->route('admin.index.post', ['store_id'=>$store->id])->with('danger', 'Удаление невозможно! Причина: присутствует '.$store->posts->count() .' записи, удалите их, потом повторите удаление.');
+        }
+        if($store->sliders()->count() >= 1) {
+         return back()->with('danger', 'Удаление невозможно! Причина: присутствует слайд, удалите слайд потом повторите удаление.');
+        }
         if($store->images) {
             foreach ($store->images as $image) {
                 Image::destroy($image->id);
@@ -110,6 +146,7 @@ class AdminController extends Controller
 
     public function index_post(Request $request)
     {
+
         $posts = Post::with('store');
         $title = 'Все записи';
         if($request->category_id) {
@@ -128,6 +165,7 @@ class AdminController extends Controller
 
     public function create_post(Request $request)
     {
+         
     	$category = Category::pluck('name', 'id');
     	$store = Store::where('confirmed', 1)->pluck('name', 'id');
     	if(isset($request->id)) {
@@ -149,6 +187,9 @@ class AdminController extends Controller
     public function delete_post(Request $request)
     {
         $post = Post::find($request->id);
+        if($post->sliders()->count() >= 1) {
+            return back()->with('danger', 'Удаление невозможно! Причина: присутствует слайд, удалите слайд потом повторите удаление.');
+        }
         if($post->images) {
             foreach ($post->images as $image) {
                 Image::destroy($image->id);
@@ -161,10 +202,12 @@ class AdminController extends Controller
 
     public function store_post(Request $request)
     {
+
     	$data = $request->except(['_token']);
     	$post = Post::updateOrCreate(['id' => $request->id], $data);
+        #$this->multiple_slide_upload($post, $request);
         $this->multiple_image_upload($post->id, $request, 'App\Post');
-    	return redirect()->route('admin.index.post');
+        return redirect()->route('admin.index.post');
     }
 
 }
